즐겨찾기 중복처리와 메시지 변경

+++ bookmark.py
@@ -22,25 +22,38 @@ def bookmark():
     proposer = id['proposer']
     date = id['date']
 
-    doc = {
-        "id": point_id,
-        "url": url,
-        "title": title,
-        "proposer": proposer,
-        "date": date
-    }
-    db.bookmark.insert_one(doc)
-    return jsonify({'result': 'success', 'msg': f'법안 {title} 저장!'})
+    bookmark_id = db.bookmark.find_one({'id': point_id})
 
+    if bookmark_id is not None:
+        msg = "이미 즐겨찾기에 저장된 법안입니다."
+    else:
+        doc = {
+            "id": point_id,
+            "url": url,
+            "title": title,
+            "proposer": proposer,
+            "date": date
+        }
+        print("doc", doc)
+        db.bookmark.insert_one(doc)
+        msg = "즐겨찾기에 저장되었습니다."
+
+    return jsonify({'result': 'success', 'msg': f'{msg}'})
 
 # 법안 즐겨찾기 삭제
 @bp.route('/api/delete_bookmark', methods=['POST'])
 def delete_bookmark():
 
     id_receive = request.form['id_give']
-    db.bookmark.delete_one({"id": id_receive})
+    bookmark_id = db.bookmark.find_one({'id': id_receive})
 
-    return jsonify({'result': 'success', 'msg': f'법안 삭제!'})
+    if bookmark_id is not None:
+        db.bookmark.delete_one({"id": id_receive})
+        msg = "법안이 삭제 되었습니다."
+    else:
+        msg = "즐겨찾기에 없는 법안입니다."
+
+    return jsonify({'result': 'success', 'msg': f'{msg}'})
